requirements.txt
alembic==1.8.1
anyio==3.6.2
bcrypt==4.0.1
cffi==1.15.1
cryptography==38.0.3
ecdsa==0.18.0
fastapi==0.85.1
idna==3.4
Mako==1.2.3
MarkupSafe==2.1.1
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.42
starlette==0.20.4
typing_extensions==4.4.0
